Restructored code into modules. Only public methods left in the main class. Fully deprecated shadowDOM and behavior handlers. Added options to skip default TEI behaviors, degut mode (some common messages are now omitted unless debug is true), and base now must be passed via the options.
